requirements.txt
snet-cli
dlib
opencv-contrib-python
grpcio
grpcio-tools
scikit-image